requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
flask-restful==0.3.10
flask-admin==1.6.1
email-validator==2.1.0
python-dotenv==1.0.0
werkzeug==2.3.7
flask-bcrypt==1.0.1
PyJWT==2.8.0
pymysql==1.1.0  # 如果使用MySQL
cryptography==41.0.4
